+++ src/tests/fake-proof.ts
@@ -9,17 +9,21 @@ import {
   Struct,
   Field,
   Proof,
+  Unconstrained,
+  Provable,
 } from 'o1js';
 import assert from 'assert';
 
 const RealProgram = ZkProgram({
   name: 'real',
+  publicOutput: UInt64,
   methods: {
     make: {
       privateInputs: [UInt64],
       async method(value: UInt64) {
         let expected = UInt64.from(34);
         value.assertEquals(expected);
+        return { publicOutput: value.add(1) };
       },
     },
   },
@@ -27,13 +31,19 @@ const RealProgram = ZkProgram({
 
 const FakeProgram = ZkProgram({
   name: 'fake',
+  publicOutput: UInt64,
   methods: {
-    make: { privateInputs: [UInt64], async method(_: UInt64) {} },
+    make: {
+      privateInputs: [UInt64],
+      async method(_: UInt64) {
+        return { publicOutput: UInt64.zero };
+      },
+    },
   },
 });
 
-class RealProof extends ZkProgram.Proof(RealProgram) {}
-const Nested = Struct({ inner: RealProof });
+class RealProof extends RealProgram.Proof {}
+class Nested extends Struct({ inner: RealProof }) {}
 
 const RecursiveProgram = ZkProgram({
   name: 'recursive',
@@ -46,9 +56,27 @@ const RecursiveProgram = ZkProgram({
     },
     verifyNested: {
       privateInputs: [Field, Nested],
-      async method(_unrelated, { inner }) {
-        inner satisfies Proof<undefined, void>;
+      async method(_unrelated, { inner }: Nested) {
+        inner satisfies Proof<undefined, UInt64>;
         inner.verify();
+      },
+    },
+    verifyInternal: {
+      privateInputs: [Unconstrained<Proof<undefined, UInt64> | undefined>],
+      async method(
+        fakeProof: Unconstrained<Proof<undefined, UInt64> | undefined>
+      ) {
+        // witness either fake proof from input, or real proof
+        let proof = await Provable.witnessAsync(RealProof, async () => {
+          let maybeFakeProof = fakeProof.get();
+          if (maybeFakeProof !== undefined) return maybeFakeProof;
+
+          let { proof } = await RealProgram.make(UInt64.from(34));
+          return proof;
+        });
+
+        proof.declare();
+        proof.verify();
       },
     },
   },
@@ -71,7 +99,7 @@ let { verificationKey: programVk } = await RecursiveProgram.compile();
 
 // proof that should be rejected
 const { proof: fakeProof } = await FakeProgram.make(UInt64.from(99999));
-const dummyProof = await RealProof.dummy(undefined, undefined, 0);
+const dummyProof = await RealProof.dummy(undefined, UInt64.zero, 0);
 
 for (let proof of [fakeProof, dummyProof]) {
   // zkprogram rejects proof
@@ -115,11 +143,10 @@ for (let proof of [fakeProof, dummyProof]) {
   }, 'recursive program rejects fake proof (nested)');
 }
 
+// zkprogram accepts proof (nested)
 const { proof: recursiveProofNested } = await RecursiveProgram.verifyNested(
   Field(0),
-  {
-    inner: realProof,
-  }
+  { inner: realProof }
 );
 assert(
   await verify(recursiveProofNested, programVk),
@@ -127,3 +154,23 @@ assert(
 );
 
 console.log('fake proof test passed for nested proofs 🎉');
+
+// same test for internal proofs
+
+for (let proof of [fakeProof, dummyProof]) {
+  // zkprogram rejects proof (internal)
+  await assert.rejects(async () => {
+    await RecursiveProgram.verifyInternal(Unconstrained.from(proof));
+  }, 'recursive program rejects fake proof (internal)');
+}
+
+// zkprogram accepts proof (internal)
+const { proof: internalProof } = await RecursiveProgram.verifyInternal(
+  Unconstrained.from(undefined)
+);
+assert(
+  await verify(internalProof, programVk),
+  'recursive program accepts internal proof'
+);
+
+console.log('fake proof test passed for internal proofs 🎉');